Renaming all files from bt to bullet
