+++ moveit_core/collision_detection_bullet/include/moveit/collision_detection_bullet/tesseract/bullet_discrete_bvh_manager.h
@@ -38,8 +38,9 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-#ifndef TESSERACT_COLLISION_BULLET_DISCRETE_BVH_MANAGERS_H
-#define TESSERACT_COLLISION_BULLET_DISCRETE_BVH_MANAGERS_H
+
+#ifndef MOVEIT_COLLISION_DETECTION_BULLET_TESSERACT_COLLISION_BULLET_DISCRETE_BVH_MANAGER_H_
+#define MOVEIT_COLLISION_DETECTION_BULLET_TESSERACT_COLLISION_BULLET_DISCRETE_BVH_MANAGER_H_
 
 #include <moveit/collision_detection_bullet/tesseract/bullet_utils.h>
 #include <moveit/macros/class_forward.h>
