+++ moveit_core/collision_detection_bullet/test/test_bullet_collision_detection_panda.cpp
@@ -36,8 +36,8 @@
 
 #include <moveit/robot_model/robot_model.h>
 #include <moveit/robot_state/robot_state.h>
-#include <moveit/collision_detection_bullet/collision_world_bt.h>
-#include <moveit/collision_detection_bullet/collision_robot_bt.h>
+#include <moveit/collision_detection_bullet/collision_world_bullet.h>
+#include <moveit/collision_detection_bullet/collision_robot_bullet.h>
 #include <moveit/collision_detection_fcl/collision_robot_fcl.h>
 #include <moveit/collision_detection_fcl/collision_world_fcl.h>
 #include <moveit/utils/robot_model_test_utils.h>
